Refactored Task
Refactored the class Task.
